mv build_config to Client
